Understand python, fix problem
The real problem was that yara is contained in a submodule. When installing
with pip as a github dependency, it was not automatically pulling
submodules. Now the build process understands to do that. I don't think
it's necessary to build before install but I have to fix other things
now.